+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/process/test/PreliminaryScenarioTestDataSerDe.scala
@@ -4,10 +4,11 @@ import cats.data.NonEmptyList
 import io.circe.parser
 import pl.touk.nussknacker.engine.definition.test.{PreliminaryScenarioTestData, PreliminaryScenarioTestRecord}
 import pl.touk.nussknacker.ui.api.TestDataSettings
+import pl.touk.nussknacker.ui.process.test.PreliminaryScenarioTestDataSerDe.{DeserializationError, SerializationError}
 
 class PreliminaryScenarioTestDataSerDe(testDataSettings: TestDataSettings) {
 
-  def serialize(scenarioTestData: PreliminaryScenarioTestData): Either[String, RawScenarioTestData] = {
+  def serialize(scenarioTestData: PreliminaryScenarioTestData): Either[SerializationError, RawScenarioTestData] = {
     import io.circe.syntax._
 
     val content = scenarioTestData.testRecords
@@ -17,31 +18,57 @@ class PreliminaryScenarioTestDataSerDe(testDataSettings: TestDataSettings) {
     Either.cond(
       content.length <= testDataSettings.testDataMaxLength,
       RawScenarioTestData(content),
-      s"Too much data generated, limit is: ${testDataSettings.testDataMaxLength}"
+      SerializationError.TooManyCharactersGenerated(length = content.length, limit = testDataSettings.testDataMaxLength)
     )
   }
 
-  def deserialize(rawTestData: RawScenarioTestData): Either[String, PreliminaryScenarioTestData] = {
+  def deserialize(rawTestData: RawScenarioTestData): Either[DeserializationError, PreliminaryScenarioTestData] = {
     import cats.implicits.catsStdInstancesForEither
     import cats.syntax.either._
     import cats.syntax.traverse._
 
-    val rawRecords = rawTestData.content.linesIterator.toList
     for {
+      _ <- Either.cond(
+        rawTestData.content.length <= testDataSettings.testDataMaxLength,
+        (),
+        DeserializationError.TooManyCharacters(
+          length = rawTestData.content.length,
+          limit = testDataSettings.testDataMaxLength
+        )
+      )
+      rawRecords = rawTestData.content.linesIterator.toList
       _ <- Either.cond(
         rawRecords.size <= testDataSettings.maxSamplesCount,
         (),
-        s"Too many samples: ${rawRecords.size}, limit is: ${testDataSettings.maxSamplesCount}"
+        DeserializationError.TooManySamples(size = rawRecords.size, limit = testDataSettings.maxSamplesCount)
       )
       decodedRecords <- rawRecords.map { rawTestRecord =>
-        val record = parser.decode[PreliminaryScenarioTestRecord](rawTestRecord)
-        record.leftMap(_ => s"Could not parse record: '$rawTestRecord'")
+        val parsedRecord = parser.decode[PreliminaryScenarioTestRecord](rawTestRecord)
+        parsedRecord.leftMap(_ => DeserializationError.RecordParsingError(rawTestRecord))
       }.sequence
       result <- NonEmptyList
         .fromList(decodedRecords)
         .map(nel => Right(PreliminaryScenarioTestData(nel)))
-        .getOrElse(Left("Empty list of records"))
+        .getOrElse(Left(DeserializationError.NoRecords))
     } yield result
+  }
+
+}
+
+object PreliminaryScenarioTestDataSerDe {
+  sealed trait SerializationError
+
+  object SerializationError {
+    final case class TooManyCharactersGenerated(length: Int, limit: Int) extends SerializationError
+  }
+
+  sealed trait DeserializationError
+
+  object DeserializationError {
+    final case class TooManyCharacters(length: Int, limit: Int) extends DeserializationError
+    final case class TooManySamples(size: Int, limit: Int)      extends DeserializationError
+    final case class RecordParsingError(rawTestRecord: String)  extends DeserializationError
+    final case object NoRecords                                 extends DeserializationError
   }
 
 }